pyproject.toml
[project]
name = "calibration-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Alexandre Lemiere", email = "alexandre.lemiere11@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "ace-tools>=0.0",
    "arff>=0.9",
    "fsspec>=2025.5.1",
    "huggingface-hub>=0.33.2",
    "ipykernel>=6.29.5",
    "kagglehub>=0.3.12",
    "marimo[lsp,recommended]>=0.14.11",
    "matplotlib>=3.10.3",
    "node>=1.2.2",
    "numpy>=2.3.1",
    "pre-commit>=4.2.0",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.7.0",
    "seaborn>=0.13.2",
    "uncertainty-calibration>=0.1.4",
    "venn-abers>=1.4.6",
    "watchdog>=6.0.0",
    "xgboost>=3.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
